+++ test/unit_test/manual_dpki/notification.js
@@ -1,4 +1,4 @@
-const { simple_conductor_config, simple_2_conductor_config, handleHack } = require('../config')
+const { simple_conductor_config, simple_2_conductor_config, handleHack } = require('../../config')
 
 async function liza_conductor_init (agent){
   return await agent.call('dpki_happ', "dpki", "init_dpki",  {params: "{\"revocation_key\": \"HcSCiPdMkst9geux7y7kPoVx3W54Ebwkk6fFWjH9V6oIbqi77H4i9qGXRsDcdbi\",\"signed_auth_key\":\"zJkRXrrbvbzbH96SpapO5lDWoElpzB1rDE+4zbo/VthM/mp9qNKaVsGiVKnHkqT4f5J4MGN+q18xP/hwQUKyDA==\"}"})
